isolate knowledge about batons to binding.cpp
